Remove CSS pane.
The CSS pane was showing the source of the library, but the intent
of Styleboard is to document the API of the patterns.  The expectation
was that the CSS there was required when instantiating the pattern,
which it wasn't.

We also get more room for everything else by shifting the splits.